Fix casting errorr in the javas

+++ canton-etf-api/src/main/java/com/canton/etf/api/dto/CreateRebalanceRequest.java
@@ -1,10 +1,11 @@
 package com.canton.etf.api.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CreateRebalanceRequest(
         String proposalId,
         List<WeightEntry> newWeights
 ) {
-    public record WeightEntry(String symbol, double weight) {}
+    public record WeightEntry(String symbol, BigDecimal weight) {}
 }
